UPSTREAM: Smack: Privilege check on key operations
Smack: Privilege check on key operations

Operations on key objects are subjected to Smack policy
even if the process is privileged. This is inconsistent
with the general behavior of Smack and may cause issues
with authentication by privileged daemons. This patch
allows processes with CAP_MAC_OVERRIDE to access keys
even if the Smack rules indicate otherwise.

Reported-by: Jose Bollo <jobol@nonadev.net>
Change-Id: Iae2abe5d2ab3b234184240c2d4e9f8a2b7fca58e
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>