pgo: Fix the profiler locking mechanism
Based on testing and earlier prototypes it is an very bad idea
to block in the profiler hook.
All such attempts will eventually dead lock the system.

So change the profiler hook locking to be try-lock based.

The prf_lock() is replaced with the new prf_lock_exclusive()
machinery that uses atomic flag and synchronize_rcu() to disable
profile data updates while the pgo_mutex is held.

Since this is a normal sleepable mutex
the serialization code is simplified.

The old pgo_lock spin lock will be removed entirely after we get
the per-cpu profiler machinery going.

Signed-off-by: Jarmo Tiitto <jarmo.tiitto@gmail.com>